// src/modules/cards/enums/card-type.enum.ts
export enum CardType {
  Virtual = 'virtual',
  Physical = 'physical',
}

export enum HolderAccountType {
  Personal = 'personal',
  Business = 'business',
}

export enum HolderVersion {
  V1 = 'v1',
  V2 = 'v2',
}

export enum DocumentType {
  Passport = 'passport',
  NationalId = 'national_id',
  DriversLicense = 'drivers_license',
  HkHkid = 'hk_hkid',
}